+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## [1.4.8](https://github.com/RedHatInsights/ocp-advisor-frontend/compare/v1.4.7...v1.4.8) (2022-06-14)
+
+## [1.4.7](https://github.com/RedHatInsights/ocp-advisor-frontend/compare/v1.4.6...v1.4.7) (2022-06-13)
+
+
+### Bug Fixes
+
+* **Recommendation clusters:** Fix always checked checkbox in the disable rule modal ([#316](https://github.com/RedHatInsights/ocp-advisor-frontend/issues/316)) ([4e4231e](https://github.com/RedHatInsights/ocp-advisor-frontend/commit/4e4231eef4421cc0f483a4f4cb623b5f87778b53))
+
+## [1.4.6](https://github.com/RedHatInsights/ocp-advisor-frontend/compare/v1.4.5...v1.4.6) (2022-06-10)
+
+
+### Bug Fixes
+
+* **Cluster reports:** fix infinite table re-render when there are no rules ([#315](https://github.com/RedHatInsights/ocp-advisor-frontend/issues/315)) ([54632ef](https://github.com/RedHatInsights/ocp-advisor-frontend/commit/54632efa79593f5e2a4df2219e8526f60d8585ca))
+
 ## [1.4.5](https://github.com/RedHatInsights/ocp-advisor-frontend/compare/v1.4.4...v1.4.5) (2022-06-07)